jancumps: removed the imports.mak[.OS_name]
Fix from @jancumps, it cleans up the imports.mak which was causing issues for those that install the SDK via CCS, because that SDK is bundled slightly differently and does not include an imports.mak file. Also, it was discovered that the imports.mak file didn't really do a lot, so jancumps was able to get rid of its requirement.